feat: adapt test-infra to refactoring from node
